Remove Unittest code from Function App.
