--- AzureProjectGraderFunctionApp/AzureProjectGrader/Constants.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace AzureProjectGrader
-{
-    public class Constants
-    {
-        public static readonly string ResourceGroupName = "projProd";
-        public static readonly string Vnet1Name = "projVnet1Prod";
-        public static readonly string Vnet2Name = "projVnet2Prod";
-
-    }
-}
